requirements/requirements.txt
numpy==1.17.0
opencv-python==4.1.0.25
PyAudio-0.2.11-cp37-cp37m-win_amd64.whl